# Speedup for `did2s()`