--- a/R/did2s.R
+++ b/R/did2s.R
@@ -153,7 +153,7 @@ did2s <- function(data, yname, first_stage, second_stage, treatment, cluster_var
     } else {
       weights_vector <- sqrt(data[[weights]])
     }
-    
+
     # Extract first stage
     first_u <- est$first_u
     if (!is.null(removed_rows)) first_u <- first_u[removed_rows]
@@ -183,7 +183,7 @@ did2s <- function(data, yname, first_stage, second_stage, treatment, cluster_var
     # Note: SparseM relies on A (x10'x10) being positive symmetric for solving
     V <- MatrixExtra::t_deep(
       SparseM::solve(
-        Matrix::crossprod(x10), 
+        Matrix::crossprod(x10),
         MatrixExtra::t_shallow(Matrix::crossprod(x2, x1))
       )
     )
@@ -195,17 +195,19 @@ did2s <- function(data, yname, first_stage, second_stage, treatment, cluster_var
     # W_g = X_2g' e_2g - (X_2' X_12) (X_11' X_11)^-1 X_11g' e_1g
     #     = X_2g' e_2g - V X_11g' e_1g
     meat <- lapply(unique(cl), function(cl_id) {
-      x2_g = make_g(x2, cl, cl_id)
-      x10_g = make_g(x10, cl, cl_id)
-      first_u_g = make_g(first_u, cl, cl_id)
-      second_u_g = make_g(second_u, cl, cl_id)
+      in_cl = (cl == cl_id)
+      n_row = sum(in_cl)
+      x2_g = make_g(x2, in_cl, n_row)
+      x10_g = make_g(x10, in_cl, n_row)
+      first_u_g = make_g(first_u, in_cl, n_row)
+      second_u_g = make_g(second_u, in_cl, n_row)
 
       W = Matrix::crossprod(x2_g, second_u_g) - V %*% Matrix::crossprod(x10_g, first_u_g)
-      
+
       # W' W
       Matrix::tcrossprod(W)
     })
-    
+
     meat_sum <- Reduce("+", meat)
 
     # (X_2'X_2)^-1 (sum W_g W_g') (X_2'X_2)^-1
@@ -335,22 +337,20 @@ did2s_estimate <- function(data, yname, first_stage, second_stage, treatment,
 
 # Subset sparse matrices
 # This is needed when subset has 1 column or 1 row
-make_g <- function(x, cl, cl_id) {
-  in_cl <- (cl == cl_id)
+make_g <- function(x, in_cl, n_row) {
   ncol <- dim(x)[[2]]
-  nrow <- sum(in_cl)
 
   if (inherits(x, "dgCMatrix")) {
-    if (nrow == 1 | ncol == 1) {
+    if (n_row == 1 | ncol == 1) {
       return(
-        Matrix::Matrix(x[cl == cl_id, ],
+        Matrix::Matrix(x[in_cl, ],
           sparse = T,
-          nrow = nrow, ncol = ncol
+          nrow = n_row, ncol = ncol
         )
       )
     } else {
       return(
-        Matrix::Matrix(x[cl == cl_id, ], sparse = T)
+        Matrix::Matrix(x[in_cl, ], sparse = T)
       )
     }
   }