@kylebutts 2023-03-29
Thanks! I was putting off doing that b/c I think I know how to speed that up by even more; but actual progress is much better than theoretical progress! 

Fair point on having better tests. At this point I know the point estimates on the `df_hom` estimator, so I don't personally need them, but that's not sustainable. I'll add some tests here in a bit