@kylebutts 2023-03-29
Thanks for the inspiration, I sped it up some more (and much lower memory) [96da875](https://github.com/kylebutts/did2s/commit/96da875a5e61e0c105c55f7ec590f1b47ec29469). Tests added too [475b749](https://github.com/kylebutts/did2s/commit/475b7492a2164c523909624b017b145b8554e74d)

It is far too easy to accidentally make a sparse matrix dense... 🙃

```r
library(data.table)
library(did2s)

foo <- list()
for (i in 1:2500) {
  g <- paste0("Group ", sample(1:3, 1))
  start_rel <- sample(c((-20):(-1), Inf), 1)
  if (is.infinite(start_rel)) {
    rel_year <- rep(Inf, 31)
  } else {
    rel_year <- seq(from = start_rel, by = 1, length.out = 31)
  }

  foo[[i]] <- data.frame(
    unit = rep(i, 31),
    state = rep(i + 10000, 31),
    group = rep(g, 31),
    rel_year = rel_year,
    dep_var = rnorm(31),
    year = 1990:2020
  )
  foo[[i]]$treat <- as.numeric(!is.infinite(foo[[i]]$rel_year) & foo[[i]]$rel_year > 0)
}

dat <- data.table::rbindlist(foo)

bench::mark(
  did2s = did2s(dat,
    yname = "dep_var", first_stage = ~ 0 | state + year,
    second_stage = ~ i(treat, ref = FALSE), treatment = "treat",
    cluster_var = "state"
  )
)

# Your fixes:
# expression      min   median `itr/sec` mem_alloc 
# did2s         3.88s    3.88s     0.257    4.04GB

# Update:
# expression      min   median `itr/sec` mem_alloc 
# did2s          2.8s     2.8s     0.357     438MB
```