@etiennebacher 2023-03-29
Wow, what an improvement in memory alloc, nice!